+++ src/Service/Models/Requests/ProofOfWorkRequest.cs
@@ -2,7 +2,7 @@
 
 public record ProofOfWorkRequest
 {
-    public int? TotalHashRate { get; init; }
+    public double? TotalHashRate { get; init; }
     
     public int? SecondsUntilBlock { get; init; }
 }
